requirements.txt
flask==3.0.2
werkzeug==3.0.1
opencv-python==4.9.0.80
numpy==1.26.4
Pillow==10.2.0
moviepy==1.0.3
gunicorn==21.2.0 